requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
python-multipart==0.0.9
sqlalchemy==2.0.27
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
plotly==5.18.0
statsmodels==0.14.1
prophet==1.1.5
requests==2.31.0
pymongo==4.6.1
cryptography==42.0.5
pyjwt==2.8.0
pydantic==2.6.1
pytest==7.4.4
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
mypy==1.8.0
bandit==1.7.6
flake8==7.0.0
